+++ lib/scandir/scandir.py
@@ -35,15 +35,14 @@ try:
 except ImportError:
     ctypes = None
 
-if not getattr(_scandir, 'DirEntry', None):
+
+if None is not _scandir and None is not ctypes and not getattr(_scandir, 'DirEntry', None):
     import warnings
-    warnings.warn("scandir compiled _scandir C module is too old"
-                  ", using slow generic fallback")
+    warnings.warn("scandir compiled _scandir C module is too old, using slow generic fallback")
     _scandir = None
 elif _scandir is None and ctypes is None:
     import warnings
-    warnings.warn("scandir can't find the compiled _scandir C module "
-                  "or ctypes, using slow generic fallback")
+    warnings.warn("scandir can't find the compiled _scandir C module or ctypes, using slow generic fallback")
 
 __version__ = '1.9.0'
 __all__ = ['scandir', 'walk']